slight fixes

+++ fnv.h
@@ -2,6 +2,7 @@
 #define FNV_H
 
 #include <cstdint>
+#include <iostream>
 #include <string>
 
 const uint64_t FNV_prime_64 = 1099511628211;
@@ -11,7 +12,7 @@ uint64_t fnv1_64(const std::string &str, const int &funcIndex) {
   uint64_t hash_value = offset_basis_64;
 
   for (char c : str + std::to_string(funcIndex)) {
-    hash_value = (hash_value * FNV_prime_64) ^ static_cast<uint8_t>(c);
+    hash_value = (hash_value * FNV_prime_64) ^ static_cast<int>(c);
   }
 
   return hash_value;
